frontend/requirements.txt
flask
gunicorn
geojson
beautifulsoup4
requests
google-cloud-secret-manager
google-cloud-storage